PubMiner/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pubminer"
version = "0.1.0"
description = "Intelligent medical literature mining and structured analysis tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "PubMiner Team"}
]
keywords = ["pubmed", "literature-mining", "llm", "medical-research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "aiohttp>=3.9.0",
    "biopython>=1.81",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pandas>=2.0.0",
    "zhipuai>=2.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
pubminer = "pubminer.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pubminer*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
